Rename SHTC1 functions to distinguish SHTC3 from others
Certain functions of the SHTC1 driver only work with SHTC3 sensors. To
have a better idea of what function works with which sensor, we change
the function name prefix to `shtcx_` for all functions that work for all
sensors of the SHTC1 family (SHTC1, SHTW2, SHTC3) and to `shtc3_` for
all functions that only work with an SHTC3.